+++ Website/Sources/Components/SectionHeader.swift
@@ -1,13 +1,13 @@
 import Ignite
 
 struct SectionHeader: HTML {
-  let id: String
-  let title: String
+  let type: SectionType
+  let language: Language
 
   var body: some HTML {
     ZStack(alignment: .center) {
       Text(markdown: "\n---")
-      Text(title)
+      Text(String(forKey: type.rawValue, language: language))
         .horizontalAlignment(.center)
         .font(.title1)
         .fontWeight(.bold)
@@ -15,6 +15,6 @@ struct SectionHeader: HTML {
     }
     .padding(.top, 80)
     .padding(.bottom, 16)
-    .id(id)
+    .id(type.rawValue)
   }
 }
